chore: refactor oc-table-cell-data component using script setup
