--- packages/design-system/src/components/_OcTableCellData/_OcTableCellData.vue
@@ -1,45 +1,0 @@
-<template>
-  <oc-table-cell
-    type="td"
-    :align-h="alignH"
-    :align-v="alignV"
-    :width="width"
-    :wrap="wrap"
-    class="oc-td"
-  >
-    <slot />
-  </oc-table-cell>
-</template>
-<script lang="ts">
-import { defineComponent } from 'vue'
-import OcTableCell from '../OcTableCell/OcTableCell.vue'
-
-export default defineComponent({
-  name: 'OcTd',
-  status: 'ready',
-  release: '2.1.0',
-  components: { OcTableCell },
-  props: {
-    alignH: {
-      type: String,
-      default: 'left',
-      validator: (alignment: string) => ['left', 'center', 'right'].includes(alignment)
-    },
-    alignV: {
-      type: String,
-      default: 'middle',
-      validator: (alignment: string) => ['top', 'middle', 'bottom'].includes(alignment)
-    },
-    width: {
-      type: String,
-      default: 'auto',
-      validator: (width: string) => ['auto', 'shrink', 'expand'].includes(width)
-    },
-    wrap: {
-      type: String,
-      default: null,
-      validator: (wrap: string) => ['break', 'nowrap', 'truncate'].includes(wrap) || !wrap
-    }
-  }
-})
-</script>
