+++ packages/design-system/src/components/index.ts
@@ -41,7 +41,7 @@ export { default as OcSwitch } from './OcSwitch/OcSwitch.vue'
 export { default as OcTable } from './OcTable/OcTable.vue'
 export { default as OcTableCell } from './OcTableCell/OcTableCell.vue'
 export { default as OcTableBody } from './OcTableBody/OcTableBody.vue'
-export { default as OcTableCellData } from './_OcTableCellData/_OcTableCellData.vue'
+export { default as OcTableCellData } from './OcTableCellData/OcTableCellData.vue'
 export { default as OcTableCellHead } from './_OcTableCellHead/_OcTableCellHead.vue'
 export { default as OcTableFooter } from './_OcTableFooter/_OcTableFooter.vue'
 export { default as OcTableHeader } from './_OcTableHeader/_OcTableHeader.vue'
